Create SideMapControlDateRange component, adjust font size
