+++ atd-vzv/package.json
@@ -15,7 +15,6 @@
     "@fortawesome/free-solid-svg-icons": "^5.12.0",
     "@fortawesome/react-fontawesome": "^0.1.8",
     "@material-ui/core": "^4.7.1",
-    "@wojtekmaj/react-daterange-picker": "^2.5.0",
     "axios": "^0.19.0",
     "bootstrap": "^4.3.1",
     "chart.js": "^2.8.0",
